+++ out/configurationHelper.js
@@ -53,6 +53,7 @@ const defaultConfig = {
     finalNewLine: 'sameAsSourceFile',
     darkThemeTextColor: '#d0d0d0',
     lightThemeTextColor: '#657b83',
+    convertUrlsToTags: true,
 };
 /**
  * returns the configuration for this extension