+++ out/configurationHelper.js
@@ -53,7 +53,7 @@ const defaultConfig = {
     finalNewLine: 'sameAsSourceFile',
     darkThemeTextColor: '#d0d0d0',
     lightThemeTextColor: '#657b83',
-    convertUrlsToTags: true,
+    convertUrlsToLinkTags: true,
 };
 /**
  * returns the configuration for this extension