+++ src/configurationHelper.ts
@@ -53,7 +53,7 @@ const defaultConfig: EditCsvConfig = {
 	finalNewLine: 'sameAsSourceFile',
 	darkThemeTextColor: '#d0d0d0',
 	lightThemeTextColor: '#657b83',
-	convertUrlsToTags: true,
+	convertUrlsToLinkTags: true,
 }
 
 /**